pack: 패키지 조정
