+++ src/main/java/org/jun/saemangeum/process/application/collect/api/GimjeCultureCollector.java
@@ -4,8 +4,7 @@ import org.jun.saemangeum.process.application.collect.base.OpenApiCollector;
 import org.jun.saemangeum.process.application.util.TitleDuplicateChecker;
 import org.jun.saemangeum.process.domain.dto.RefinedDataDTO;
 import org.jun.saemangeum.process.infrastructure.api.OpenApiClient;
-import org.jun.saemangeum.process.presentation.dto.GimjeCultureResponse;
-import org.jun.saemangeum.process.presentation.dto.GunsanCultureResponse;
+import org.jun.saemangeum.process.application.dto.GimjeCultureResponse;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
