+++ src/main/java/org/jun/saemangeum/process/application/dto/Festival.java
@@ -1,9 +1,7 @@
-package org.jun.saemangeum.process.presentation.dto;
+package org.jun.saemangeum.process.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
-import org.jun.saemangeum.global.domain.Category;
-import org.jun.saemangeum.process.domain.dto.RefinedDataDTO;
 
 @Getter
 public class Festival {
