+++ src/main/java/org/jun/saemangeum/process/application/dto/FestivalResponse.java
@@ -1,13 +1,13 @@
-package org.jun.saemangeum.process.presentation.dto;
+package org.jun.saemangeum.process.application.dto;
 
 import java.util.List;
 
-public record EventResponse(
+public record FestivalResponse(
         int page,
         int perPage,
         int totalCount,
         int currentCount,
         int matchCount,
-        List<Event> data
+        List<Festival> data
 ) {
 }
