+++ src/main/java/org/jun/saemangeum/process/application/dto/GunsanCultureResponse.java
@@ -1,13 +1,13 @@
-package org.jun.saemangeum.process.presentation.dto;
+package org.jun.saemangeum.process.application.dto;
 
 import java.util.List;
 
-public record FestivalResponse(
+public record GunsanCultureResponse(
         int page,
         int perPage,
         int totalCount,
         int currentCount,
         int matchCount,
-        List<Festival> data
+        List<GunsanCulture> data
 ) {
 }
