+++ src/main/java/org/jun/saemangeum/process/presentation/TestController.java
@@ -2,7 +2,6 @@ package org.jun.saemangeum.process.presentation;
 
 import lombok.RequiredArgsConstructor;
 import org.jun.saemangeum.process.application.service.ContentDataProcessService;
-import org.jun.saemangeum.process.presentation.dto.TestDTO;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
